Collect all results.
Brought in a lot of narrative, but also threw a lot out.  Took out
everything dealing with creating incidence lists since that is really
all previous work and our implementation does not really provide anything
new.  Also took out discussion of local averaging.  This made more sense
with the original bottom up discussion and local averaging explenation.
But with the new narritive it is just confusing.

I also added placeholders for results with the cell subdivision and
face-centered tetrahedralization, which were missing.